requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.3
openpyxl==3.1.5
requests==2.32.3
sqlalchemy==2.0.41
psycopg[binary]==3.2.6
pydantic==2.11.10
pydantic-settings==2.9.1
python-dotenv==1.1.0
orjson==3.10.18
